exercise/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
Django==4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
pytz==2022.1
sqlparse==0.4.2
